Updated scatter plot to include a legend and updated my API calls. I also added Commits
